Worked on ray example .. added video save.  Fixed dtype=float issue with ray image creation.  Fixed inv bug in homography inverse transformation.

+++ software/studio50_fly/studio50_fly/utility.py
@@ -37,7 +37,8 @@ def create_ray_image(x0, y0, angle, image_shape, num_rays, color=(0,255,0)):
         theta1 = (2.0*np.pi*((2*i+1)/float(2*num_rays))) 
         mask_arc = np.logical_and(coord_angles >= theta0, coord_angles < theta1)
         mask = np.logical_or(mask, mask_arc)
-    ray_image = np.zeros((nrow, ncol, nchan), dtype=np.float)
+    #ray_image = np.zeros((nrow, ncol, nchan), dtype=np.float)
+    ray_image = np.zeros((nrow, ncol, nchan), dtype=np.uint8)
     ray_image[mask,:] = color
     return ray_image
